[PATCH] Bug#24947436 UPGRADE RAPIDJSON TO VERSION 1.1.0
Upgrade the rapidjson library to version 1.1.0. This version
contains fixes that remove compiler warnings on newer versions
of clang, plus several other enhancements and bugfixes.

We previously had a modification in the library, where we could
choose to handle "-0" as a double value in order to preserve
the signedness. This was mainly used by GIS code. Now, we send a
flag to the JSON parser telling us whether to parse ALL numbers
as double or not. This effectively uses the RawNumber callback in
rapidjson instead of the individual Int64/Double/Int... callbacks.
We now use this functionality instead of modifying the library,
so that all custom modifications can be removed.

Change-Id: I784ad30bca474ba1872ebb4e2d5867c4e11b9dd2